Added pagination to search controller. Renamed to follow convention. Moved support functions from dataset controller into dataset service. Updated documentation. Refactored findAll to be consistent for GET and POST.
