+++ api/app/controllers/search.controller.js
@@ -2,19 +2,27 @@ const strategyManager = require('../services/strategyManager');
 const datasetsService = require("../services/datasetsService");
 
 async function search(req, res) {
-    const {query, sequence} = req.body;
-    console.log('running search', query, sequence);
+    const { query, sequence, page, rows, limit, filters, nofacets } = req.body;
+    console.log('running search', query, sequence, page, rows, limit, filters, nofacets);
 
     try {
         if (sequence && sequence.trim() !== '') {
             console.log('retrieving sequence results', sequence);
             const results = await strategyManager.runSearch(query, sequence);
-            res.json(results);
-        } else {
-            console.log('retrieving local results');
-            const localResults = await datasetsService.searchLocalDatasets({textQueryTerm: query});
-            res.json(localResults);
+            return res.json(results);
         }
+
+        console.log('retrieving local results');
+        const localResults = await datasetsService.searchLocalDatasets({
+          textQueryTerm: query,
+          page,
+          rows,
+          limit,
+          filters,
+          nofacets,
+        });
+
+        return res.json(localResults);
     } catch (error) {
         console.error('Error during search:', error);
         res.status(500).json({message: 'Search failed. Please try again later.'});
